Move identity provider implementations into kbase namespace.
