+++ src/us/kbase/test/auth2/kbase/GoogleIdentityProviderTest.java
@@ -1,4 +1,4 @@
-package us.kbase.test.auth2.lib.identity;
+package us.kbase.test.auth2.kbase;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
@@ -27,14 +27,14 @@ import org.mockserver.model.ParameterBody;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
+import us.kbase.auth2.kbase.GoogleIdentityProvider;
+import us.kbase.auth2.kbase.GoogleIdentityProvider.GoogleIdentityProviderConfigurator;
 import us.kbase.auth2.lib.exceptions.IdentityRetrievalException;
-import us.kbase.auth2.lib.identity.GoogleIdentityProvider;
 import us.kbase.auth2.lib.identity.IdentityProvider;
 import us.kbase.auth2.lib.identity.IdentityProviderConfig;
 import us.kbase.auth2.lib.identity.RemoteIdentity;
 import us.kbase.auth2.lib.identity.RemoteIdentityDetails;
 import us.kbase.auth2.lib.identity.RemoteIdentityID;
-import us.kbase.auth2.lib.identity.GoogleIdentityProvider.GoogleIdentityProviderConfigurator;
 import us.kbase.auth2.lib.identity.IdentityProviderConfig.IdentityProviderConfigurationException;
 import us.kbase.test.auth2.TestCommon;
